Replaced BoltPluginLoaders and BoltDirectoryManagers with BoltLoaders.
While a BoltPluginLoader/BoltDirectoryManager combo was responsible for all the ways that a type of plugin could be discovered, a BoltLoader is 1 way to load 1 type of plugin. For example, loading Java plugins from jars is handled by a different loader than loading Java plugins from class objects. This allows BoltPluginManager to be more generic, rather than subclasses providing a single classloader-plugin-loader and a single filesystem-plugin-loader, it now accepts any number of BoltLoaders and runs through them in sequence.