BoltJarContainers don't count builtins
BoltJarContainer was using a URLClassLoader to load plugins from a jar,
but the ServiceLoader reading it was also seeing any builtins from the
classloader's parent classloader. A check is now performed to make sure
any plugin classes registered by a BoltJarContainer were loaded by its
URLClassLoader and not a parent.