Split EscapePeaks opt into ShowEscapePeaks and DetectorMaterial
This closes #124 'Turning off escape peaks has bizarre side-effects on
peak width'.

It turns out that the detector material specified by the escape peak
property is used for more than just escape peaks. It is also used to
calculate peak widths. Setting the escape peaks to none zeroes out some
parts of that calculation.

Splitting this option in to and removing the None material solved the
issue.