PlotFrame can prompt for confirmation on exit
This is the start of work on #173

PlotFrame will now poll each PlotPanel and call its new hasUnsavedWork
method. If any of them do, it will prompt the user.

What's missing is an implementation for the hasUnsavedWork method. Right
now it's just returning false. The undo stack may provide some basis for
making this determination, but it's unclear how.