Drag-and-drop for datasets and plugins off UI thread, shows progress bar
Plural's Streams classes made a good base for a new stipped down
interface TaskMonitor. From there, SimpleTaskMonitor just handles a
simple task with Supplier<T> to do work and Consumer<T> to be called
when it's done. The function setPercent allows the Supplier to update
the progress as it goes.