Map Selection impls expose param Group, controller is more passthrough
The Selection implementations return an AutoDialog Group of Parameters
so that the UI can automatically construct the right settings components
for the current selection tool.

The Selection interfacfe adds some of the methods alread in
MapSelectionController so that it can just call the current selection
impl rather than having a bunch of switch statements, since the switch
statements defeat the purpose of the OO design.