Non-rectangular/contiguous selections can be replotted/mapped w/ limits
The limits are that the re-mapped data cannot be re-re-plotted, because
the re-mapped maps contain points for which there is no valid spectrum
associated with it. Even without that, its going to take some more work
before it will be able to trace re-mapped points back to the original
spectra.