Further filter/user map size mismatch mitigation
The issue appears to be (at least partly) related to the event broadcast
system. When the map dimensions change, it emits an event which goes to
several listeners. One of those listeners is the filtering controller,
which will invalidate the filters after that point, another is the UI
canvas element which will repaint the window. Depending on the order,
the repaint could happen before the filter in validation. This change
delays the repaint by submitting it as another task on the swing event
queue.

This still doesn't seem to completely resolve the issue, and I'm not
sure why.