Eventful UI Queue Priority Fix
There was a bug/problem where sometimes the handling of one message
would prompt another component to emith another message. The message
chain would normally propagate correctly, unless there was additional
input processed on the UI thread between the messages. When that
happened, the state of the earlier components might not match what the
later-messaged components were expecting.

The solution is to dispatch all eventfull messages to the UI thread via
a single job queue. With each submission to the queue, a job will also
be submitted to the UI thread which will run each job in the queue until
the queue is empty. This means that if the first message causes more
messages to be emitted, then those messages will be added to the glboal
job queue before the first delivery job on the UI thread finishes.

This allows whole chains of messages to be delivered in one shot,
without the chance of other UI events being processed in between.

This does, finally, seem to resolve and close #218