Updated PlainText/CSV parser to handle multiple spaces
With other delimiters like ',' having two of them suggests that there
is a zero/null value there. With spaces, it's possible to have it repeat
without that, eg:

0        1        2

The parser was seeing each space as a separator and inserting 0 values
between them. This has been changed to leaving them as null values and
dealing with them differently at the point of parsing each value/number
based on the delimiter detected.