Fix for hover button background drawing bug, further LaF refinements
Buttons which only draw their background when 'acrtive' (eg hover) were
drawing a full rectangle of the wrong colour behind the button. This
appears to be unrelated to the Stratus drawing code. A new way of
indicating if the button's border should be drawn is used that does not
interact with any other swing components or methods.